% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_accessibility.R
\name{aggregate_accessibility}
\alias{aggregate_accessibility}
\alias{aggregate_accessibility.gtfs_accessibility}
\title{Aggregate Accessibility}
\usage{
aggregate_accessibility(x, ...)

\method{aggregate_accessibility}{gtfs_accessibility}(x, districts, ...)
}
\arguments{
\item{x}{A gtfs_accessibility object.}

\item{...}{Additional arguments passed to methods.}

\item{districts}{An sf polygon layer for districts or a path to a vector file.}
}
\value{
A gtfs_district_accessibility object containing aggregated
accessibility statistics for each district.
}
\description{
Aggregates accessibility to larger districts polygons.
}
\examples{
\dontrun{
districts <- system.file(
  "extdata",
  "districts.geojson",
  package = "gtfsaccess"
)

district_accessibility <-
  aggregate_accessibility(
    accessibility,
    districts
  )
}
}
