% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_stops.R
\name{calculate_geojson_area}
\alias{calculate_geojson_area}
\title{Calculate GeoJSON Area}
\usage{
calculate_geojson_area(geojson_file)
}
\arguments{
\item{geojson_file}{Path to a GeoJSON file.}
}
\value{
Area in square kilometres.
}
\description{
Calculates the study area in square kilometres.
}
\examples{
\dontrun{
calculate_geojson_area("study_area.geojson")
}
}
