% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_stops.R
\name{calculate_stop_density}
\alias{calculate_stop_density}
\title{Calculate Stop Density}
\usage{
calculate_stop_density(number_of_stops, area_km2)
}
\arguments{
\item{number_of_stops}{Number of stops.}

\item{area_km2}{Study area in square kilometres.}
}
\value{
A list of stop density statistics.
}
\description{
Calculates stop density statistics.
}
\examples{
\dontrun{
density <- calculate_stop_density(
  number_of_stops = 500,
  area_km2 = 100
)

density
}
}
