% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_accessibility.R
\name{count_gtfs_departures}
\alias{count_gtfs_departures}
\title{Count GTFS Departures}
\usage{
count_gtfs_departures(gtfs_zip)
}
\arguments{
\item{gtfs_zip}{Path to a GTFS ZIP file.}
}
\value{
A data frame containing stop_id and departures.
}
\description{
Counts the total number of departures for each stop in a GTFS feed.
}
\examples{
\dontrun{
departures <- count_gtfs_departures("muenster_gtfs.zip")

head(departures)
}
}
