% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_stops.R
\name{count_gtfs_stops}
\alias{count_gtfs_stops}
\title{Count GTFS Stops}
\usage{
count_gtfs_stops(stops)
}
\arguments{
\item{stops}{A GTFS stops data frame.}
}
\value{
Number of unique stops.
}
\description{
Counts the number of unique GTFS stops.
}
\examples{
\dontrun{
stops <- read_gtfs_stops("muenster_gtfs.zip")

count_gtfs_stops(stops$stops)
}
}
