% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_accessibility.R
\name{extract_population}
\alias{extract_population}
\title{Extract Population by Catchment}
\usage{
extract_population(service_areas, population)
}
\arguments{
\item{service_areas}{An sf object.}

\item{population}{A SpatRaster.}
}
\value{
The service areas with a population_in_catchment column.
}
\description{
Extract Population by Catchment
}
\examples{
\dontrun{
population <- terra::rast("population.tif")

service_areas <- create_service_areas(
  read_gtfs_stops("muenster_gtfs.zip"),
  "boundary.geojson"
)

service_areas <- extract_population(
  service_areas,
  population
)

head(service_areas)
}
}
