% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_accessibility.R
\name{join_departures}
\alias{join_departures}
\title{Join Departures}
\usage{
join_departures(service_areas, departures)
}
\arguments{
\item{service_areas}{An sf object.}

\item{departures}{A data frame.}
}
\value{
Updated service areas.
}
\description{
Join Departures
}
\examples{
\dontrun{
departures <- count_gtfs_departures("muenster_gtfs.zip")

service_areas <- create_service_areas(
  read_gtfs_stops("muenster_gtfs.zip"),
  "boundary.geojson"
)

service_areas <- join_departures(
  service_areas,
  departures
)

head(service_areas)
}
}
