% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_gtfs_accessibility}
\alias{new_gtfs_accessibility}
\title{Create a GTFS Accessibility Object}
\usage{
new_gtfs_accessibility(accessibility)
}
\arguments{
\item{accessibility}{An sf object containing accessibility results.}
}
\value{
An object of class \code{gtfs_accessibility}.
}
\description{
Creates an S3 object representing accessibility results
derived from GTFS service areas and population data.
}
\examples{
\dontrun{
library(sf)

accessibility <- st_sf(
  stop_id = c("S1", "S2"),
  population_in_catchment = c(1250, 890),
  departures = c(320, 180),
  accessibility_index = c(0.256, 0.202),
  geometry = st_sfc(
    st_polygon(list(rbind(
      c(0,0), c(1,0), c(1,1), c(0,1), c(0,0)
    ))),
    st_polygon(list(rbind(
      c(1,0), c(2,0), c(2,1), c(1,1), c(1,0)
    )))
  ),
  crs = 4326
)


obj <- new_gtfs_accessibility(
  accessibility,
)

print(obj)
summary(obj)
}
}
