% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_gtfs_stops}
\alias{new_gtfs_stops}
\title{Create a GTFS Stops Object}
\usage{
new_gtfs_stops(stops)
}
\arguments{
\item{stops}{An sf object containing GTFS stop locations.}
}
\value{
An object of class \code{gtfs_stops}.
}
\description{
Creates an S3 object representing GTFS stop locations.
}
\examples{
\dontrun{
library(sf)

stops <- data.frame(
  stop_id = c("S1", "S2"),
  stop_name = c("Central Station", "University"),
  stop_lon = c(7.6261, 7.5952),
  stop_lat = c(51.9607, 51.9694)
)

stops_sf <- st_as_sf(
  stops,
  coords = c("stop_lon", "stop_lat"),
  crs = 4326
)

obj <- new_gtfs_stops(stops_sf)

print(obj)
summary(obj)
}
}
