% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.gtfs_stops}
\alias{plot.gtfs_stops}
\title{Plot GTFS Stops}
\usage{
\method{plot}{gtfs_stops}(x, pch = 16, col = "red", cex = 0.7, ...)
}
\arguments{
\item{x}{A gtfs_stops object.}

\item{pch}{Point symbol.}

\item{col}{Point colour.}

\item{cex}{Point size.}

\item{...}{Additional graphical arguments.}
}
\description{
Plot GTFS Stops
}
