% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.gtfs_accessibility}
\alias{print.gtfs_accessibility}
\title{Print a GTFS Accessibility Object}
\usage{
\method{print}{gtfs_accessibility}(x, ...)
}
\arguments{
\item{x}{A gtfs_accessibility object.}

\item{...}{Additional arguments.}
}
\description{
Print a GTFS Accessibility Object
}
