% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.gtfs_district_accessibility}
\alias{print.gtfs_district_accessibility}
\title{Print a GTFS District Accessibility Object}
\usage{
\method{print}{gtfs_district_accessibility}(x, ...)
}
\arguments{
\item{x}{A gtfs_district_accessibility object.}

\item{...}{Additional arguments.}
}
\description{
Print a GTFS District Accessibility Object
}
