% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.gtfs_stops}
\alias{print.gtfs_stops}
\title{Print a GTFS Stops Object}
\usage{
\method{print}{gtfs_stops}(x, ...)
}
\arguments{
\item{x}{A gtfs_stops object.}

\item{...}{Additional arguments.}
}
\description{
Print a GTFS Stops Object
}
