% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_stops.R
\name{read_gtfs_stops}
\alias{read_gtfs_stops}
\title{Read GTFS Stops}
\usage{
read_gtfs_stops(gtfs_zip)
}
\arguments{
\item{gtfs_zip}{Path to a GTFS ZIP file.}
}
\value{
A \code{gtfs_stops} object.
}
\description{
Reads a GTFS feed and returns the stop locations as a
\code{gtfs_stops} object.
}
\examples{
\dontrun{
stops <- read_gtfs_stops("muenster_gtfs.zip")

print(stops)
summary(stops)
plot(stops)
}
}
