% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.gtfs_accessibility}
\alias{summary.gtfs_accessibility}
\title{Summarize a GTFS Accessibility Object}
\usage{
\method{summary}{gtfs_accessibility}(object, ...)
}
\arguments{
\item{object}{A gtfs_accessibility object.}

\item{...}{Additional arguments.}
}
\description{
Summarize a GTFS Accessibility Object
}
