% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.gtfs_stops}
\alias{summary.gtfs_stops}
\title{Summarize a GTFS Stops Object}
\usage{
\method{summary}{gtfs_stops}(object, ...)
}
\arguments{
\item{object}{A gtfs_stops object.}

\item{...}{Additional arguments.}
}
\description{
Summarize a GTFS Stops Object
}
